Updated logic for Near Wallet & Ledger Wallet.
