+++ src/wallets/hardware/LedgerWallet.ts
@@ -1,7 +1,6 @@
 import { transactions, utils } from "near-api-js";
 import { TypedError } from "near-api-js/lib/utils/errors";
 import isMobile from "is-mobile";
-import BN from "bn.js";
 import ProviderService from "../../services/provider/ProviderService";
 import { Emitter } from "../../utils/EventsHandler";
 import LedgerClient, { Subscription } from "./LedgerClient";
@@ -11,12 +10,12 @@ import { setSelectedWalletId } from "../helpers";
 import { ledgerWalletIcon } from "../icons";
 import {
   AccountInfo,
-  FunctionCallAction,
   HardwareWallet,
   HardwareWalletType,
   SignAndSendTransactionParams,
   WalletOptions,
 } from "../Wallet";
+import { transformActions } from "../Actions";
 
 interface AuthData {
   accountId: string;
@@ -236,17 +235,6 @@ class LedgerWallet implements HardwareWallet {
     };
   };
 
-  private transformActions = (actions: Array<FunctionCallAction>) => {
-    return actions.map((action) => {
-      return transactions.functionCall(
-        action.methodName,
-        action.args,
-        new BN(action.gas),
-        new BN(action.deposit)
-      );
-    });
-  };
-
   signAndSendTransaction = async ({
     receiverId,
     actions,
@@ -273,7 +261,7 @@ class LedgerWallet implements HardwareWallet {
       utils.PublicKey.from(publicKey),
       receiverId,
       accessKey.nonce + 1,
-      this.transformActions(actions),
+      transformActions(actions),
       utils.serialize.base_decode(block.header.hash)
     );
